+++ jgltf-model/src/main/java/de/javagl/jgltf/model/animation/StepInterpolator.java
@@ -33,7 +33,7 @@ class StepInterpolator implements Interpolator
 {
     @Override
     public void interpolate(
-        float[] a, float[] b, float alpha, float[] result)
+        double[] a, double[] b, double alpha, double[] result)
     {
         System.arraycopy(a, 0, result, 0, a.length);
     }
